first test finish ok

+++ GarciaJuradoJuanManuel_pruebatec4/GarciaJuradoJuanManuel_pruebatec4/pruebaTecnica/src/main/java/AgenciaTurismo/pruebaTecnica/service/HotelService.java
@@ -3,8 +3,10 @@ package AgenciaTurismo.pruebaTecnica.service;
 import AgenciaTurismo.pruebaTecnica.dto.NewHotelDTO;
 import AgenciaTurismo.pruebaTecnica.model.Hotel;
 import AgenciaTurismo.pruebaTecnica.model.Room;
+import AgenciaTurismo.pruebaTecnica.model.RoomReservation;
 import AgenciaTurismo.pruebaTecnica.repository.HotelRepository;
 import AgenciaTurismo.pruebaTecnica.repository.RoomRepository;
+import AgenciaTurismo.pruebaTecnica.repository.RoomReservationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,6 +21,8 @@ public class HotelService implements IHotelService  {
     private HotelRepository hotelRepo;
     @Autowired
     private RoomRepository roomRepo;
+    @Autowired
+    private RoomReservationRepository roomReservationRepo;
 
     @Override
     public void saveHotel(NewHotelDTO newHotel) {
@@ -39,19 +43,24 @@ public class HotelService implements IHotelService  {
 
     @Override
     public void deleteHotel(Long id) {
-        // Buscar el hotel por su ID
+        // Buscar el hotel por su ID. Find the hotel by its ID
         Hotel hotelToDelete = hotelRepo.findById(id).orElse(null);
 
         if (hotelToDelete != null) {
-            // Actualizar las relaciones de las habitaciones asociadas a este hotel
+            // Verificar si hay reservas de habitaciones asociadas con el hotel. Check if there are room reservations associated with the hotel
+            List<RoomReservation> reservations = roomReservationRepo.findByRoom_Hotel_Id(id);
+            if (!reservations.isEmpty()) {
+                throw new RuntimeException("Cannot delete hotel with associated room reservations");
+            }
+
+            // Actualizar las relaciones de las habitaciones asociadas a este hotel. Update the relationships of the rooms associated with this hotel
             List<Room> rooms = roomRepo.findByHotelId(id);
             for (Room room : rooms) {
-                // Por ejemplo, puedes establecer el ID del hotel asociado a NULL
                 room.setHotel(null);
                 roomRepo.save(room);
             }
 
-            // Luego de actualizar las relaciones, eliminar el hotel
+            // Luego de actualizar las relaciones, eliminar el hotel. After updating the relationships, delete the hotel
             hotelRepo.deleteById(id);
         }
     }
